rewape the variables name files name and alignment the code structure
