+++ package.json
@@ -4,9 +4,9 @@
   "description": "user registration curd",
   "main": "index.js",
   "scripts": {
-    "build": "babel source -s -D -d library",
-    "start": "npm run build && node --env-file=.env library/index.js",
-    "development": "nodemon --exec babel-node --env-file=.env source/index.js",
+    "build": "babel src -s -D -d lib",
+    "start": "npm run build && node --env-file=.env lib/index.js",
+    "development": "nodemon --exec babel-node --env-file=.env src/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "prashanth",
