+++ src/utlis/sendMail.js
@@ -1,4 +1,4 @@
-import nodemailer from "nodemailer";
+import nodemailer from 'nodemailer';
 import dotenv from 'dotenv';
 
 dotenv.config();
