+++ src/validators/auth-validator.js
@@ -1,5 +1,5 @@
-import Joi from "joi";
-import * as response from "../utlis/httputlis";
+import Joi from 'joi';
+import * as response from '../utlis/httputlis';
 
 const register = Joi.object({
     name: Joi.string().min(3).max(50).required(),
@@ -16,11 +16,11 @@ export  const registerValidation = (req , res , next) => {
     try {
         const {error} = register.validate(req.body);
         if (error) {
-            return response.errorResponse(res, response.HttpStatus.BAD_REQUEST,"Validation failed", { error: error.details[0].message })
+            return response.errorResponse(res, response.HttpStatus.BAD_REQUEST,'Validation failed', { error: error.details[0].message })
         }
         next();
     } catch (error) {
-      return response.errorResponse(res, response.HttpStatus.INTERNAL_SERVER_ERROR, "Internal Server Error", { error: error.message });
+      return response.errorResponse(res, response.HttpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', { error: error.message });
     }
 }
 
@@ -28,10 +28,10 @@ export const loginValidation = (req , res , next) => {
     try {
         const {error} = login.validate(req.body);
         if (error) {
-            return response.errorResponse(res, response.HttpStatus.BAD_REQUEST,"Validation failed", { error: error.details[0].message })
+            return response.errorResponse(res, response.HttpStatus.BAD_REQUEST,'Validation failed', { error: error.details[0].message })
         }
         next();
     } catch (error) {
-      return response.errorResponse(res, response.HttpStatus.INTERNAL_SERVER_ERROR, "Internal Server Error", { error: error.message });
+      return response.errorResponse(res, response.HttpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', { error: error.message });
     }
 }
