+++ src/validators/subadmin-validator.js
@@ -1,5 +1,5 @@
-import Joi from "joi";
-import * as response from "../utlis/httputlis";
+import Joi from 'joi';
+import * as response from '../utlis/httputlis';
 
 const subAdminCreate = Joi.object({
     name: Joi.string().min(3).max(50).required(),
@@ -12,10 +12,10 @@ export const subAdminValidation = (req,res,next) => {
     try {
         const {error} = subAdminCreate.validate(req.body);
         if (error) {
-            return response.errorResponse(res, response.HttpStatus.BAD_REQUEST,"Validation failed", { error: error.details[0].message })
+            return response.errorResponse(res, response.HttpStatus.BAD_REQUEST,'Validation failed', { error: error.details[0].message })
         }
         next();
     } catch (error) {
-      return response.errorResponse(res, response.HttpStatus.INTERNAL_SERVER_ERROR, "Internal Server Error", { error: error.message });
+      return response.errorResponse(res, response.HttpStatus.INTERNAL_SERVER_ERROR, 'Internal Server Error', { error: error.message });
     }
 }
